requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
aniso8601==9.0.1
annotated-types==0.6.0
anyio==4.3.0
async-timeout==4.0.3
attrs==24.2.0
blinker==1.7.0
blis==0.7.11
cachelib==0.9.0
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
Cython==0.29.37
dataclasses-json==0.6.7
distro==1.9.0
elastic-transport==8.12.0
elasticsearch==8.12.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
exceptiongroup==1.2.0
faiss-cpu==1.9.0
filelock==3.13.1
Flask==3.0.2
Flask-Caching==2.3.0
Flask-Cors==4.0.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
fonttools==4.48.1
frozenlist==1.4.1
fsspec==2024.2.0
gensim==4.3.2
gradio_client==0.15.0
groq==0.11.0
h11==0.14.0
hdbscan==0.8.33
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.26.2
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
jiter==0.6.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.5
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-groq==0.2.0
langchain-openai==0.2.2
langchain-text-splitters==0.3.0
langcodes==3.3.0
langsmith==0.1.134
llvmlite==0.42.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.8.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.10
mypy-extensions==1.0.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.4
openai==1.51.2
orjson==3.10.7
packaging==23.2
pandas==2.2.0
pillow==10.2.0
plotly==5.18.0
preshed==3.0.9
propcache==0.2.0
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pynndescent==0.5.12
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.1.1
regex==2023.12.25
requests==2.31.0
requests-toolbelt==1.0.0
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.10.1
sentence-transformers==3.2.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
spacy==3.7.2
spacy-alignments==0.9.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-transformers==1.3.5
SQLAlchemy==2.0.27
srsly==2.4.8
sympy==1.12
tenacity==8.2.3
thinc==8.2.3
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.20.2
top2vec==1.0.33
torch==2.2.1
tqdm==4.66.2
transformers==4.46.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
urllib3==1.26.18
wasabi==1.1.2
weasel==0.3.4
websockets==11.0.3
Werkzeug==3.0.1
wordcloud==1.9.3
yarl==1.15.1
zipp==3.17.0
